öğrenmek için denemeler
