+++ app/src/main/java/com/example/iot_obdii/Data.java
@@ -1,0 +1,6 @@
+package com.example.iot_obdii;
+
+public class Data {
+    String key;
+    String value;
+}
