+++ app/src/main/res/layout/activity_main.xml
@@ -13,7 +13,6 @@ tools:context=".MainActivity">
         android:layout_height="wrap_content"
         android:layout_alignParentLeft="true"
         android:layout_alignParentTop="true"
-        android:onClick="connectBTN"
         android:text="Start" />
 
     <TextView
